refactor code to allow multiple images uploading in upload_multi_images, and separate logic from upload_photo_dialog
